+++ DocCodeSamples.Tests/LocalizedStringDatabaseSamples.cs
@@ -1,4 +1,5 @@
 using System.Collections;
+using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.AddressableAssets;
 using UnityEngine.Localization.Settings;
@@ -52,3 +53,119 @@ public class ReleaseSample : MonoBehaviour
 }
 
 #endregion
+
+public class GetLocalizedStringExamples
+{
+    public void GetLocalizedStringAsyncExample1()
+    {
+        #region get-string-async
+
+        var stringOperation = LocalizationSettings.StringDatabase.GetLocalizedStringAsync("My String");
+        stringOperation.Completed += s =>
+        {
+            Debug.Log("String loaded: " + s);
+        };
+        #endregion
+    }
+
+    public void GetLocalizedStringAsyncExample2()
+    {
+        #region get-string-async-2
+
+        var stringOperation = LocalizationSettings.StringDatabase.GetLocalizedStringAsync("My Table", "My String");
+        stringOperation.Completed += s =>
+        {
+            Debug.Log("String loaded: " + s);
+        };
+        #endregion
+    }
+
+    public void GetLocalizedStringAsyncWithArgumentExample()
+    {
+        #region get-string-argument-async
+
+        // Example string: "The value is {0}"
+        var stringOperation = LocalizationSettings.StringDatabase.GetLocalizedStringAsync("My String", new object[] { 123 });
+        stringOperation.Completed += s =>
+        {
+            // Example output: "The value is 123"
+            Debug.Log(s.Result);
+        };
+        #endregion
+    }
+
+    public void GetLocalizedStringAsyncWithSmartArgumentExample()
+    {
+        #region get-string-smart-argument-async
+
+        // Prepare Smart String Arguments.
+        var dictionary = new Dictionary<string, string>();
+        dictionary.Add("title", "General");
+        dictionary.Add("name", "Radahn");
+
+        // Example string: "I am {title} {name}!"
+        var stringOperation = LocalizationSettings.StringDatabase.GetLocalizedStringAsync("Main Characters", "General_Radahn_Intro", new object[] { dictionary });
+        stringOperation.Completed += s =>
+        {
+            // Example output: "I am General Radahn!"
+            Debug.Log(s.Result);
+        };
+        #endregion
+    }
+
+    public void GetLocalizedStringNoFallbackExample()
+    {
+        #region get-string-smart-argument
+
+        // Prepare Smart String Arguments.
+        var dictionary = new Dictionary<string, string>();
+        dictionary.Add("title", "General");
+        dictionary.Add("name", "Radahn");
+
+        // Example string: "I am {title} {name}!"
+        var localizedString = LocalizationSettings.StringDatabase.GetLocalizedString("Main Characters", "General_Radahn_Intro", new object[] { dictionary });
+
+        // Example output: "I am General Radahn!"
+        Debug.Log(localizedString);
+        #endregion
+    }
+
+    #region get-string-coroutine-async
+
+    IEnumerator LoadStringWithCoroutine()
+    {
+        var stringOperation = LocalizationSettings.StringDatabase.GetLocalizedStringAsync("My String");
+
+        // We can skip waiting if the operation is already done.
+        if (!stringOperation.IsDone)
+            yield return stringOperation;
+
+        Debug.Log(stringOperation.Result);
+    }
+    #endregion
+
+    public void GetLocalizedStringAsyncWaitForCompletion()
+    {
+        #region get-string-custom-locale-async
+
+        var customLocale = LocalizationSettings.AvailableLocales.GetLocale("fr");
+
+        var stringOperation = LocalizationSettings.StringDatabase.GetLocalizedStringAsync("My String", customLocale);
+
+        // We can force the operation to complete.
+        if (!stringOperation.IsDone)
+            stringOperation.WaitForCompletion();
+
+        Debug.Log(stringOperation.Result);
+        #endregion
+    }
+
+    public void GetLocalizedStringExample()
+    {
+        #region get-string
+
+        var localizedString = LocalizationSettings.StringDatabase.GetLocalizedString("My Table", "My String");
+        Debug.Log(localizedString);
+        #endregion
+    }
+}