+++ Editor/UI/Tables/StringTableListView.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using UnityEditor.IMGUI.Controls;
 using UnityEngine;
 using UnityEngine.Localization.Tables;
@@ -24,6 +25,8 @@ namespace UnityEditor.Localization.UI
             });
             base.UndoRedoPerformed();
         }
+
+        protected override IReadOnlyCollection<StringTable> GetCollectionTables() => (TableCollection as StringTableCollection).StringTables;
 
         protected override float GetCustomRowHeight(int row, TreeViewItem item)
         {